requirements.txt
django-jinja
markdown
pyyaml
